requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
mysql-connector-python>=8.0.0

